2001-08-08 23:52 GMT+2 Maurilio Longo <maurilio.longo@libero.it>    * include/hbvmpub.h    * source/vm/classes.c    * source/vm/hvm.c    * source/vm/dynsym.c      ! Fixed profiler counting of time elapsed inside recursive functions. It still has to        be fixed for methods.        NOTE: My knowledge of harbour VM is very poor. I hope I didn't break anything else :-))      ! Fixed a few warnings inside hb_vmDo() and hb_vmSend()    * source/vm/arrayshb.c      ! Fixed isalnum() call on line 359, isalnum() requires an int

+++ harbour/source/vm/arrayshb.c
@@ -356,7 +356,7 @@ HB_FUNC( HB_AEXPRESSIONS )
             break;
 
          case '[' :
-            if( bArray || ( i && isalnum( pLine->item.asString.value[i - 1] ) ) )
+            if( bArray || ( i && isalnum((int) pLine->item.asString.value[i - 1] ) ) )
             {
                iIndexs++;
             }
